flask_backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0
elevenlabs~=1.3.0
deepgram-sdk==2.11.0
groq==0.4.1
numpy==1.25.2
pandas==2.1.0
openpyxl==3.1.2